Remove Microsoft.CodeAnalysis.CSharp reference
This analyzer is intended for code using the Roslyn compiler API. The specs project that referenced it doesn't appear to be using that, so it looks like this was added in error.

The packages.lock.json updates here are because it would appear that dependabot doesn't update those, so although the csproj files had the relevant updates, the lock files did not.